requirements.txt
transformers
torch
langfuse
requests
python-dotenv